mavlink: move GLOBAL_POSITION_INT to separate stream header
