+++ src/modules/mavlink/streams/GLOBAL_POSITION_INT.hpp
@@ -1,0 +1,124 @@
+/****************************************************************************
+ *
+ *   Copyright (c) 2021 PX4 Development Team. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the
+ *    distribution.
+ * 3. Neither the name PX4 nor the names of its contributors may be
+ *    used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
+ * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ ****************************************************************************/
+
+#ifndef GLOBAL_POSITION_INT_HPP
+#define GLOBAL_POSITION_INT_HPP
+
+#include <uORB/topics/home_position.h>
+#include <uORB/topics/vehicle_air_data.h>
+#include <uORB/topics/vehicle_global_position.h>
+#include <uORB/topics/vehicle_local_position.h>
+
+class MavlinkStreamGlobalPositionInt : public MavlinkStream
+{
+public:
+	static MavlinkStream *new_instance(Mavlink *mavlink) { return new MavlinkStreamGlobalPositionInt(mavlink); }
+
+	static constexpr const char *get_name_static() { return "GLOBAL_POSITION_INT"; }
+	static constexpr uint16_t get_id_static() { return MAVLINK_MSG_ID_GLOBAL_POSITION_INT; }
+
+	const char *get_name() const override { return get_name_static(); }
+	uint16_t get_id() override { return get_id_static(); }
+
+	unsigned get_size() override
+	{
+		return _gpos_sub.advertised() ? MAVLINK_MSG_ID_GLOBAL_POSITION_INT_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES : 0;
+	}
+
+private:
+	explicit MavlinkStreamGlobalPositionInt(Mavlink *mavlink) : MavlinkStream(mavlink) {}
+
+	uORB::Subscription _gpos_sub{ORB_ID(vehicle_global_position)};
+	uORB::Subscription _lpos_sub{ORB_ID(vehicle_local_position)};
+	uORB::Subscription _home_sub{ORB_ID(home_position)};
+	uORB::Subscription _air_data_sub{ORB_ID(vehicle_air_data)};
+
+	bool send() override
+	{
+		vehicle_global_position_s gpos;
+		vehicle_local_position_s lpos;
+
+		if (_gpos_sub.update(&gpos) && _lpos_sub.update(&lpos)) {
+
+			mavlink_global_position_int_t msg{};
+
+			if (lpos.z_valid && lpos.z_global) {
+				msg.alt = (-lpos.z + lpos.ref_alt) * 1000.0f;
+
+			} else {
+				// fall back to baro altitude
+				vehicle_air_data_s air_data{};
+				_air_data_sub.copy(&air_data);
+
+				if (air_data.timestamp > 0) {
+					msg.alt = air_data.baro_alt_meter * 1000.0f;
+				}
+			}
+
+			home_position_s home{};
+			_home_sub.copy(&home);
+
+			if ((home.timestamp > 0) && home.valid_alt) {
+				if (lpos.z_valid) {
+					msg.relative_alt = -(lpos.z - home.z) * 1000.0f;
+
+				} else {
+					msg.relative_alt = msg.alt - (home.alt * 1000.0f);
+				}
+
+			} else {
+				if (lpos.z_valid) {
+					msg.relative_alt = -lpos.z * 1000.0f;
+				}
+			}
+
+			msg.time_boot_ms = gpos.timestamp / 1000;
+			msg.lat = gpos.lat * 1e7;
+			msg.lon = gpos.lon * 1e7;
+
+			msg.vx = lpos.vx * 100.0f;
+			msg.vy = lpos.vy * 100.0f;
+			msg.vz = lpos.vz * 100.0f;
+
+			msg.hdg = math::degrees(wrap_2pi(lpos.heading)) * 100.0f;
+
+			mavlink_msg_global_position_int_send_struct(_mavlink->get_channel(), &msg);
+
+			return true;
+		}
+
+		return false;
+	}
+};
+
+#endif // GLOBAL_POSITION_INT_HPP
